+++ Tests/StreamChatTests/Models/Attachments/ImageAttachmentPayload_Tests.swift
@@ -32,7 +32,6 @@ final class ImageAttachmentPayload_Tests: XCTestCase {
         // Assert values are decoded correctly.
         XCTAssertEqual(payload.title, title)
         XCTAssertEqual(payload.imageURL, imageURL)
-        XCTAssertEqual(payload.imagePreviewURL, thumbURL)
         XCTAssertEqual(payload.originalWidth, originalWidth)
         XCTAssertEqual(payload.originalHeight, originalHeight)
         XCTAssertNil(payload.extraData)
@@ -65,7 +64,6 @@ final class ImageAttachmentPayload_Tests: XCTestCase {
         // Assert values are decoded correctly.
         XCTAssertEqual(payload.title, title)
         XCTAssertEqual(payload.imageURL, imageURL)
-        XCTAssertEqual(payload.imagePreviewURL, thumbURL)
         
         // Assert extra data can be decoded.
         let extraData = try XCTUnwrap(payload.extraData(ofType: ExtraData.self))
@@ -76,7 +74,6 @@ final class ImageAttachmentPayload_Tests: XCTestCase {
         let payload = ImageAttachmentPayload(
             title: "Image1.png",
             imageRemoteURL: URL(string: "dummyURL")!,
-            imagePreviewRemoteURL: URL(string: "dummyPreviewURL"),
             originalWidth: 100,
             originalHeight: 50,
             extraData: ["isVerified": true]